+++ src/lib/sunburst/node.ts
